NeuralStyleTransfer_WebApp/requirements.txt
numpy==1.26.4
Pillow==10.0.1
Pillow==10.3.0
streamlit==1.29.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1